topology: Introduce new class keyword
Topology today has some common classes that are often reused
throughout with slightly altered configurations
i.e. widgets (components), pipelines, dais and controls.

This PR introduces the high level concept of reusable "class" like
definitions that can be used to create topology objects e.g.
Class.Component - Class for widgets that can be included within
Class.Pipeline. Class definitions typically contain arguments
and attributes. It also adds support for creating a new
class topology element and saving it to the class_list. This
will be used as the template when creating objects of this
class type.

A class typically consists of arguments and attributes. Arguments
and attributes both refer to parameters associated with the class
or its object. The difference is that the arguments are also used
for naming the object.

For ex:, a simple class definition would be

Class.Base."data" {

	DefineArgument."name" {
		type	"string"
	}

	DefineArgument."index" {}

	DefineAttribute."bytes" {}
}

The argument "name" would be used to build the name of the data
object as "data.<name>.<index>".

When parsing the class definition, the arguments and attributes
are stored as part of the attribute_list in struct tplg_class.
Ad attribute/argument can have values that can be of type
integer, long, double or a string.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>