topology: class: Add token_ref field
Attributes/arguments in a class could refer to tuple
values that need to be added to the private data.
Add a field called token_ref to struct tplg_attribute
stores the name of SectionVendorTokens and the tuple type
that will be used to build the tuple value for the attribute.
For ex: "sof_tkn_dai.word" refers to the SectionVendorTokens
with the name "sof_tkn_dai" and "word" refers to the tuple
type.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>