topology: class: add constraints for attributes
One of the main advantages with Topology2.0 is the
ability to validate the types and values for
the arguments and attributes for a class instance.
Add the constraint field for struct tplg_attribute that
will be used to parse and save the constraints
specified for the attribute/argument in the class definition.

Typical constraints include min/max value or a set
of valid values allowed for the attribute. The set
of valid values can also be translated from a string value
to an integer value during topology build.

An example for attribute constraint would be:
DefineAttribute."direction" {
		token_ref	"sof_tkn_dai.word"
		constraints {
			value_ref       "sof_tkn_direction"
			values [
				"playback"
				"capture"
			]
		}
}

where the value_ref "sof_tkn_direction" would translate the values
as follows:
SectionVendorTokens."sof_tkn_direction" {
	playback	"0"
	capture	"1"
}

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>