topology: Introduce the new object keyword
The Object keyword is used to instantiate a class.
When an object is instantiated, it must be provided
with arguments and mandatory attributes. Example
object instantiation:

	Object.host."N" {
		direction		"playback"
		period_sink_count	2
		period_source_count	0
		widget_type		"aif_in"
	}
where 'N' is a unique identifier for this object in the
parent alsaconf node. For objects that have an index
attribute, this value is copied to the index attribute
by the compiler.

The default attributes values for a class are copied
over to the object when the object is created.
But these will be overriden when they are explictly
set during object instantation as above.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>