topology: object: Inherit attributes from parent object
One of the key features of Topology2.0 is the concept
of inheriting object attribute values from parent
attribute values. Inheritance is implicit when an attribute
in a child object shares the same name as an attribute
in the parent object. When creating an object, make sure
to pass down the attribute values to all its child objects
recursively.

Alternatively, a parent object can also explictly set
an attribute for its child object by using the
object's class name and index attribute as follows:

Object.volume-playback."1" {
	pipeline_id	1
	pcm_id		0
	pcm_name	"Port0"
	format		"s24le"

	# Set value for mixer.0 in pga.0
	pga.0.mixer.0.name "1 Master Playback Volume"
}

The last line in the above object instance, sets the
mixer name for the mixer object with index 0 in the
PGA object with index 0 in the volume-playback object.

As a rule of thumb, the attribute value set for an object always
overrides inherited values from parent or the ones explicitly
set in the parent.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>