topology: object: build topology elements for all objects
Once an object is created, the next step is to create
the topology element for the topology builder. This patch
adds support for build BASE type objects specifically for
manifest and data objects.

Before building the topology elements, the compiler
needs to ensure that all the mandatory attributes
have been provided. Also, if attributes have references
for valid values, they need to be translated to the
appropriate tuple value.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>