topology: object: add support for building private data
Many topology elements pack tuple data in their private
data. Iterate through the attribute list of an object
and build tuple sets for those attributes that have the
token_ref field set. Add a new field, tuple_set_list
to struct objects and add the tuple sets to the list.

Once the tuple_sets are built for an object, the compiler
scans the tuple sets against the provided token reference
and builds the private data array.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>