topology2: update automatic attributes for host class
Attributes have a constraint called
TPLG_CLASS_ATTRIBUTE_MASK_AUTOMATIC that allows
the values to be updated by the alsatplg compiler
when the object containing the attribute is
built. One example of an automatic attribute
is the "stream_name" attribute for "host"
and "copier" class objects. Add the logic to update the
stream name for such objects.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>